requirements.txt
torch
scikit-learn
pandas
numpy
streamlit
funcyou==1.0.8
seaborn
scikit-image
Pillow
opencv-python
scipy
tqdm
xgboost
wandb